# [Performance] Qwen3-Next: optimize causal_conv1d_fn triton kernel - up to 9% faster