<!-- Thank you for your contribution! Please follow these guidelines to enhance your pull request. If anything is unclear, submit your PR and reach out to maintainers for assistance. Join our Slack community at https://slack.sglang.ai to discuss further. -->

## Motivation

(This PR fixed the merge conflict in this PR: https://github.com/sgl-project/sglang/pull/10552)

Currently `causal_conv1d_fn` triton kernel is used for extend & draft_extend. From the profile, it is inefficient - many CPU<>GPU copy sync and its `grid` method is very slow (takes 139us to calculate the triton grid) while the real triton kernel launch itself only takes 17us.

<img width="1137" height="228" alt="image" src="https://github.com/user-attachments/assets/890e92aa-6e70-4e1a-b46d-1a604b25b866" />

## Modifications
Take this request as example: 
- seq_lens = [1, 9, 2, 17, 4]
- BLOCK_M=8
- num_of_blockm_s_per_seq_len = torch.tensor([1, 2, 1, 3, 1]) -> total_blockm_s = 1+2+1+3+1=8
- m_list = [0, 1, 1, 2, 3, 3, 3, 4]: len(m_list) = 8 = total_blockm_s
- offset_list = [0, 0, 1, 0, 0, 1, 2, 0]: len(offset_list) = 8 = total_blockm_s

### Before
- Each program_id(0) in `_causal_conv1d_fwd_kernel` processes one BLOCK_M out of total 8 BLOCK_M
- Each BLOCK_M maps to one block of tokens. `program_id(0) = i` maps to `blockm[i]`, maps to (m_list[i], offset_list[i]) chunk of tokens
- For seq_lens = [1, 9, 2, 17, 4] & BLOCK_M=8, we have the following programs (aligns with `m_list = [0, 1, 1, 2, 3, 3, 3, 4]` and `offset_list = [0, 0, 1, 0, 0, 1, 2, 0]`):
-- program_id(0) = 0
-- program_id(0) = 1
-- program_id(0) = 2
-- program_id(0) = 3
-- program_id(0) = 4
-- program_id(0) = 5
-- program_id(0) = 6
-- program_id(0) = 7

### After: same processing logic as in triton_backend.py's extend_attention() method 
- Each program_id(0) in `_causal_conv1d_fwd_kernel` processes one seq_len out of total 5 seq_lens
- Each program_id(1) in `_causal_conv1d_fwd_kernel` processes one block of tokens belonging to that seq_len
- For seq_lens = [1, 9, 2, 17, 4] & BLOCK_M=8, we have the following programs (aligns with `num_of_blockm_s_per_seq_len = torch.tensor([1, 2, 1, 3, 1])`):
-- program_id(0) = 0, program_id(1) = 0
-- program_id(0) = 1, program_id(1) = 0
-- program_id(0) = 1, program_id(1) = 1
-- program_id(0) = 2, program_id(1) = 0
-- program_id(0) = 3, program_id(1) = 0
-- program_id(0) = 3, program_id(1) = 1
-- program_id(0) = 3, program_id(1) = 2
-- program_id(0) = 4, program_id(1) = 0

**By doing so, we don't have to calculate m_list & offset_list which requires cpu<> gpu copy and expensive tensor/numpy operations. Instead, we can just leverage `forward_batch.extend_seq_len_cpu` as it is in the triton kernel.** 

In addition to this main change, there are some misc changes done by this PR such as removing unnecessary code/tensor and improve the `if` condition in triton kernel

## Benchmarking and Profiling

### `causal_conv1d_fn` runtime
Before: p50=313us; avg=656us
<img width="1666" height="300" alt="image" src="https://github.com/user-attachments/assets/f931626b-2534-420c-ae63-9a0193e6e735" />

After: p50=95us; avg=108us - **3X faster**
<img width="1683" height="230" alt="image" src="https://github.com/user-attachments/assets/f7863eab-b24e-4f76-8b2a-74286ccef5cc" />

<img width="691" height="171" alt="image" src="https://github.com/user-attachments/assets/fde3520b-02a2-4477-946e-dab750776ca4" />

### GSM8k Accuracy

| Config        | Accuracy | Note   |
|---------------|----------|--------|
| TP4           | 0.955                  | BEFORE |
| TP4           | 0.950                  | AFTER |
| TP4 MTP       | 0.955           | BEFORE  |
| TP4 MTP       | 0.950              | AFTER  |
| TP4 DP2 MTP   | 0.945         |  BEFORE  |
| TP4 DP2 MTP   | 0.940        | AFTER  |

### GSM8k Output token/s

| Config      | BEFORE (tokens/s) | AFTER (tokens/s) | % Faster |
|-------------|-----------------|-----------------|----------|
| TP4         | 1613.119        | 1701.235        | +5.5%   |
| TP4 MTP     | 1888.384        | 2055.889        | +8.9%   |
| TP4 DP2 MTP | 1762.349        | 1811.621        | +2.8%   |

### Performance Comparation B/w causal_conv1d_fn Triton Kernel and Cuda Kernel
After this change, **causal_conv1d_fn triton kernel performance is comparable with its cuda kernel implementation under gsm8k!**

```
# triton
Accuracy: 0.945
Invalid: 0.000
Latency: 15.543 s
Output throughput: 2088.367 token/s

# cuda
Accuracy: 0.945
Invalid: 0.000
Latency: 15.577 s
Output throughput: 2091.005 token/s
```
## Checklist

- [x] Format your code according to the [Format code with pre-commit](https://docs.sglang.ai/developer_guide/contribution_guide.html#format-code-with-pre-commit).
- [x] Add unit tests according to the [Run and add unit tests](https://docs.sglang.ai/developer_guide/contribution_guide.html#run-and-add-unit-tests).
- [x] Update documentation according to [Write documentations](https://docs.sglang.ai/developer_guide/contribution_guide.html#write-documentations).
- [x] Provide accuracy and speed benchmark results according to [Test the accuracy](https://docs.sglang.ai/developer_guide/contribution_guide.html#test-the-accuracy) and [Benchmark the speed](https://docs.sglang.ai/developer_guide/contribution_guide.html#benchmark-the-speed).